Change datatype of wedge in mb_mode_info structure
Changed datatype of 'interintra_wedge_index', 'wedge_index'and 'wedge_sign'
from int to int8_t.

Removed the structure member 'interintra_wedge_sign' and introduced
a MACRO instead.

Observed memory footprint reduction with no impact on
encoder/decoder performance.

Stream                   Preset     Encoder   Decoder
BasketballDrill_832x480     1        ~0.2%       ~1.2%
parkrun_720p50              3        ~0.2%       ~1.4%

+++ av1/common/reconinter.h
@@ -323,8 +323,8 @@ void av1_count_overlappable_neighbors(const AV1_COMMON *cm, MACROBLOCKD *xd,
 
 void av1_init_wedge_masks();
 
-static INLINE const uint8_t *av1_get_contiguous_soft_mask(int wedge_index,
-                                                          int wedge_sign,
+static INLINE const uint8_t *av1_get_contiguous_soft_mask(int8_t wedge_index,
+                                                          int8_t wedge_sign,
                                                           BLOCK_SIZE sb_type) {
   return av1_wedge_params_lookup[sb_type].masks[wedge_sign][wedge_index];
 }
